requirements.txt
aiogram
sqlalchemy
python-dotenv
aiosqlite
requests